% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load_games}
\alias{load_games}
\title{here to Load the video games dataset}
\usage{
load_games()
}
\value{
A data frame with 20 rows and 7 variables:
\describe{
\item{title}{Name of the game}
\item{genre}{Genre of the game}
\item{platform}{Platform the game is available on}
\item{rating}{Metacritic rating out of 100}
\item{avg_playtime_hours}{Average playtime in hours}
\item{year_released}{Year the game was released}
\item{developer}{Developer of the game}
}
}
\description{
Returns the built-in dataset of 20 popular video games
including ratings, genres, playtime, and more.
}
\examples{
load_games()
}
