% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_ratings}
\alias{plot_ratings}
\title{This Plot game ratings}
\usage{
plot_ratings(n = 10)
}
\arguments{
\item{n}{Number of games to display. Default is 10.}
}
\value{
A ggplot2 bar chart.
}
\description{
Creates a bar chart of video game ratings using ggplot2.
}
\examples{
plot_ratings()
plot_ratings(n = 5)
}
