% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print.gamestats}
\alias{print.gamestats}
\title{printing method for gamestats objects}
\usage{
\method{print}{gamestats}(x, ...)
}
\arguments{
\item{x}{A gamestats S3 object.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns the object.
}
\description{
printing method for gamestats objects
}
