% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summary.gamestats}
\alias{summary.gamestats}
\title{And summary method for gamestats objects}
\usage{
\method{summary}{gamestats}(object, ...)
}
\arguments{
\item{object}{A gamestats S3 object.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns the object.
}
\description{
And summary method for gamestats objects
}
