% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{top_games}
\alias{top_games}
\title{Get top N games by rating or playtime}
\usage{
top_games(n = 5, by = "rating")
}
\arguments{
\item{n}{Number of top games to return. Default is 5.}

\item{by}{Column to rank by: "rating" or "avg_playtime_hours". Default is "rating".}
}
\value{
A data frame of the top N games.
}
\description{
Get top N games by rating or playtime
}
\examples{
top_games(5, by = "rating")
top_games(3, by = "avg_playtime_hours")
}
